requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.39.14
botocore==1.39.14
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
jmespath==1.0.1
multidict==6.6.3
pillow==11.3.0
propcache==0.3.2
py-cord==2.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
typing-extensions==4.14.1
urllib3==1.26.20
yarl==1.20.1
